# soroban-rpc: Add missing config settings in ledger entry cache on reads