--- a/cmd/soroban-rpc/internal/preflight/preflight_test.go
+++ b/cmd/soroban-rpc/internal/preflight/preflight_test.go
@@ -330,6 +330,7 @@ func TestGetPreflight(t *testing.T) {
 	params := getPreflightParameters(t, nil)
 	_, err := GetPreflight(context.Background(), params)
 	require.NoError(t, err)
+	params.LedgerEntryReadTx.Done()
 
 	// using a restarted db with caching and
 	getDB(t, true)
@@ -340,6 +341,8 @@ func TestGetPreflight(t *testing.T) {
 	params = getPreflightParameters(t, dbConfig)
 	_, err = GetPreflight(context.Background(), params)
 	require.NoError(t, err)
+	params.LedgerEntryReadTx.Done()
+	dbConfig.dbInstance.Close()
 }
 
 type benchmarkDBConfig struct {