--- a/cmd/soroban-rpc/internal/preflight/preflight_test.go
+++ b/cmd/soroban-rpc/internal/preflight/preflight_test.go
@@ -341,8 +341,8 @@ func TestGetPreflight(t *testing.T) {
 	params = getPreflightParameters(t, dbConfig)
 	_, err = GetPreflight(context.Background(), params)
 	require.NoError(t, err)
-	params.LedgerEntryReadTx.Done()
-	dbConfig.dbInstance.Close()
+	require.NoError(t, params.LedgerEntryReadTx.Done())
+	require.NoError(t, dbConfig.dbInstance.Close())
 }
 
 type benchmarkDBConfig struct {
@@ -369,10 +369,10 @@ func benchmark(b *testing.B, config benchmarkConfig) {
 		_, err := GetPreflight(context.Background(), params)
 		b.StopTimer()
 		require.NoError(b, err)
-		params.LedgerEntryReadTx.Done()
+		require.NoError(b, params.LedgerEntryReadTx.Done())
 	}
 	if dbConfig != nil {
-		dbConfig.dbInstance.Close()
+		require.NoError(b, dbConfig.dbInstance.Close())
 	}
 }